requirements.txt
# File monitoring and processing
watchdog==4.0.0            # Directory monitoring
Pillow==10.2.0             # Image processing
opencv-python==4.9.0.80    # Video processing

# AI/ML dependencies
torch==2.2.0               # PyTorch for GPU
torchvision==0.17.0        # Computer vision utilities
transformers==4.38.1       # Hugging Face models
accelerate==0.27.2         # Model optimization
bitsandbytes==0.42.0       # 8-bit quantization

# GPU monitoring
nvidia-ml-py3==7.352.0     # NVIDIA GPU monitoring (optional but recommended)

# Web interface and storage
fastapi==0.109.2           # Web API
uvicorn==0.27.1            # ASGI server
sqlalchemy==2.0.27         # Database ORM
aiosqlite==0.20.0          # Async SQLite
jinja2==3.1.3              # Templates

# Utilities
pydantic==2.6.1            # Data validation
python-dotenv==1.0.1       # Environment config
aiofiles==23.2.1           # Async file ops
psutil==5.9.8              # System monitoring 